feat(connector-function): add rayfin_kusto_v1 to the shared HelloFabric function_app.py
Per Rayfin-team feedback (function discovery requires all connector functions in
one file), relocate rayfin_kusto_v1 from a standalone HelloKusto app into the
shared HelloFabric/function_app.py, alongside rayfin_semantic_model_v1. Both
functions share the single fn.UserDataFunctions() and the lazily-created
aiohttp.ClientSession (_get_session()).

- Delete the HelloKusto/ folder.
- functions.metadata now lists BOTH rayfin_semantic_model_v1 and rayfin_kusto_v1.
- Regenerate SourceCode.zip and Deploy.zip (fabric_lib/functions.metadata carries
  both entries; function_app.py updated; other 10 zip entries byte-identical).

rayfin_kusto_v1 transforms Kusto native v1 responses into the SDK envelope
(mirrors KWE normalizeResultV1). Endpoint resolution is done backend-side
(workloads-appdev PR #1029801, commit 80b2b0b). Zips were generated locally and
should be regenerated via the Fabric portal Library Management tooling before
production.

Copilot-Session: 59bbb3df-85ef-45b3-ae05-05aa98546c3a
