+++ Templates/Python/ConnectorFunction/HelloFabric/functions.metadata
@@ -1,1 +1,1 @@
-[{"name": "rayfin_semantic_model_v1", "scriptFile": "function_app.py", "bindings": [{"name": "req", "direction": "In", "type": "httpTrigger", "methods": ["post"], "route": "rayfin_semantic_model_v1", "authLevel": "Anonymous"}], "fabricProperties": {"fabricFunctionReturnType": "StreamResponse", "fabricFunctionParameters": [{"name": "payload", "dataType": "dict"}, {"name": "accesstoken", "dataType": "str"}]}}]
+[{"name": "rayfin_semantic_model_v1", "scriptFile": "function_app.py", "bindings": [{"name": "req", "direction": "In", "type": "httpTrigger", "methods": ["post"], "route": "rayfin_semantic_model_v1", "authLevel": "Anonymous"}], "fabricProperties": {"fabricFunctionReturnType": "StreamResponse", "fabricFunctionParameters": [{"name": "payload", "dataType": "dict"}, {"name": "accesstoken", "dataType": "str"}]}}, {"name": "rayfin_kusto_v1", "scriptFile": "function_app.py", "bindings": [{"name": "req", "direction": "In", "type": "httpTrigger", "methods": ["post"], "route": "rayfin_kusto_v1", "authLevel": "Anonymous"}], "fabricProperties": {"fabricFunctionReturnType": "StreamResponse", "fabricFunctionParameters": [{"name": "payload", "dataType": "dict"}, {"name": "accesstoken", "dataType": "str"}]}}]
