--- Templates/Python/ConnectorFunction/HelloKusto/.vscode/settings.json
@@ -1,14 +1,0 @@
-{
-    "files.exclude": {
-        "obj": true,
-        "bin": true
-    },
-    "azureFunctions.deploySubpath": ".",
-    "azureFunctions.scmDoBuildDuringDeployment": true,
-    "azureFunctions.pythonVenv": ".venv",
-    "azureFunctions.projectLanguage": "Python",
-    "azureFunctions.projectRuntime": "~4",
-    "debug.internalConsoleOptions": "neverOpen",
-    "azureFunctions.projectLanguageModel": 2,
-    "azureFunctions.preDeployTask": "func: extensions install"
-}
