--- Templates/Python/ConnectorFunction/HelloKusto/functions.metadata
@@ -1,1 +1,0 @@
-[{"name": "rayfin_kusto_v1", "scriptFile": "function_app.py", "bindings": [{"name": "req", "direction": "In", "type": "httpTrigger", "methods": ["post"], "route": "rayfin_kusto_v1", "authLevel": "Anonymous"}], "fabricProperties": {"fabricFunctionReturnType": "StreamResponse", "fabricFunctionParameters": [{"name": "payload", "dataType": "dict"}, {"name": "accesstoken", "dataType": "str"}]}}]
