--- Templates/Python/ConnectorFunction/HelloKusto/host.json
@@ -1,15 +1,0 @@
-{
-  "version": "2.0",
-  "logging": {
-    "applicationInsights": {
-      "samplingSettings": {
-        "isEnabled": true,
-        "excludedTypes": "Request"
-      }
-    }
-  },
-  "extensionBundle": { 
-    "id": "Microsoft.Azure.Functions.ExtensionBundle", 
-    "version": "[4.26.2, 5.0.0)" 
-  }
-}
