--- Templates/Python/ConnectorFunction/HelloKusto/requirements.txt
@@ -1,7 +1,0 @@
-#This Requirements.txt file is auto-generated by Fabric. Do not edit manually. You can use the Library Management Tool in the portal to manage libraries.
-
-fabric-user-data-functions ~= 1.0
-requests == 2.33.1
-aiohttp==3.14.1
-asyncio==4.0.0
-azurefunctions-extensions-http-fastapi==1.0.1
